feat: refactor code into resource plugins and custom device plugins
