+++ pkg/agent/qrm-plugins/gpu/resourceplugin/interface.go
@@ -1,0 +1,36 @@
+/*
+Copyright 2022 The Katalyst Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package resourceplugin
+
+import (
+	gpuconsts "github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/gpu/consts"
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/gpu/state"
+	pluginapi "k8s.io/kubelet/pkg/apis/resourceplugin/v1alpha1"
+)
+
+// ResourcePlugin knows how to handle resource requests for a specific resource type.
+type ResourcePlugin interface {
+	ResourceName() string
+
+	GetTopologyHints(*pluginapi.ResourceRequest) (*pluginapi.ResourceHintsResponse, error)
+
+	GetTopologyAwareResources(*state.AllocationInfo) *gpuconsts.AllocatedResource
+
+	GetTopologyAwareAllocatableResources(state.GPUMap) *gpuconsts.AllocatableResource
+
+	Allocate(*pluginapi.ResourceRequest) (*pluginapi.ResourceAllocationResponse, error)
+}
